+++ src/Is.php
@@ -117,6 +117,18 @@ final class Is implements Constraint
         return new self(\array_is_list(...), 'list');
     }
 
+    /**
+     * @psalm-pure
+     *
+     * @param non-empty-string $key
+     *
+     * @return Constraint<mixed, non-empty-array<non-empty-string, mixed>>
+     */
+    public static function shape(string $key, Constraint $constraint): Constraint
+    {
+        return self::array()->and(Shape::of($key, $constraint));
+    }
+
     public function and(Constraint $constraint): Constraint
     {
         return AndConstraint::of($this, $constraint);